pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "sekimore-gw"
version = "0.0.1"
description = "AI Agent Security Gateway - DNS/Firewall/Proxy for Docker"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "sekimore-gw contributors"},
]
readme = "README.md"
dependencies = [
    # DNS Server
    "dnspython>=2.4.2",
    "dnslib>=0.9.23",

    # Configuration
    "PyYAML>=6.0.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Database
    "aiosqlite>=0.21.0",

    # Logging
    "structlog>=25.0.0",
    "python-json-logger>=2.0.0",

    # Web UI
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=12.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",

    # File monitoring
    "watchdog>=3.0.0",
]

[dependency-groups]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-describe>=2.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "httpx>=0.26.0",

    # Code Quality
    "ruff>=0.4.0",
    "mypy>=1.8.0",

    # Type Stubs
    "types-PyYAML>=6.0.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*", "describe_*"]
python_classes = ["Test*", "Describe*"]
asyncio_mode = "auto"
timeout = 30
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "e2e: End-to-end tests that require docker-compose (deselect with '-m \"not e2e\"')",
    "slow: Tests that take a long time to run",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
strict = false

[[tool.mypy.overrides]]
module = "dnslib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "netifaces.*"
ignore_missing_imports = true
